+++ CHANGELOG.md
@@ -63,7 +63,9 @@ Main (unreleased)
 
 - Added support for NS records to `discovery.dns`. (@djcode)
 
-- Improved clustering use cases for tracking GCP delta metrics in the `prometheus.exporter.gcp` (@kgeckhart) 
+- Improved clustering use cases for tracking GCP delta metrics in the `prometheus.exporter.gcp` (@kgeckhart)
+
+- Add the `targets` argument to the `prometheus.exporter.snmp` component to support passing SNMP targets at runtime. (@wildum)
 
 ### Bugfixes